requirements.txt
asgiref==3.11.1
Django==6.0.4
django-debug-toolbar==6.3.0
et_xmlfile==2.0.0
openpyxl==3.1.5
Pillow>=11.0.0
python-decouple==3.8
sqlparse==0.5.5
